[cat] support JIT instead of demos and fix some bugs

+++ include/Token.hpp
@@ -74,7 +74,7 @@ typedef enum {
     CONTINUE,
     // type
     INT,
-    FLOAT,
+    DOUBLE,
     BOOL,
     STR,
     LIST,
